PROD-3527: External Ticket Button Not Displaying on Event Page - Hosted Hublite
For imported events, it never displays the ticket button. Updated the method to check if there's a value in the return of $api->get_api_event_buy_ticket_url.
If not, I'm not updating  $args['ticket_url'], since it's initially loaded with $event->ticket_url (which is the proper value)